feat: refactor commit fetching logic into a separate function and update Home component to use it
